✨ Drastically reduce boilerplate with build_runner
You can now generate most of your boilerplate with `build_runner`. I believe this current setup also allows for the possibility of hot reload, as it returns a new object every time when asking for TypeInfo. This is not ideal for released builds, but we can look into memoization of typeinfo another time.

Next step is to have build_runner generate the Script File -> Type mapping, and warn if you attempt to export multiple types from a file.