✨ New generator for script type mapping
The new generator creates a function for mapping Godot script paths to types, instead of relying on `DartScriptLanguage.addScript`. I'm hoping that by using a function + the generator, we'll be able to have hot reload catch new scripts.

Also made it so that the editor and the game run on two different ports, so you can debug the game while the editor is running.