🐛 Fix threaded calls
Instead of having a Dart thread, we're moving the isolate to the thread that is attempting the Dart call. This isn't tested fully and I'm not 100% sure of the performance impact, but it may be better than thunking to a single thread.

Also fix a minor issue of a method using the wrong binding token.