✨ Implement Vector3 in Dart
The hope is that this is a bit more performant, as it doesn't jump to Godot every time it needs to do something, and it doesn't allocate memory for variables that are never sent to Godot.

This will be done for most of the math based Variants, but started with Vector3 to see what it was like.

refs: #8