feat: Remove requirement for 'bindingToken' .
This is a giant refactor that completely removes the need for 'bindingTokens'. This should, eventually, allow me to get ride of `.cast` except to be used as an actual dynamic cast instead of a necessity over Dart `is` / `as`.

This still needs to be tested with a more complex project to see if it actually works.